前3个页面的基本补完以及商城页面的初步开始

+++ CatMovie/app/src/main/java/com/personal/djb/catmovie/bean/movies/WaitMovieBean.java
@@ -112,7 +112,7 @@ public class WaitMovieBean {
             private String img;
             private boolean late;
             private int localPubSt;
-            private int mk;
+            private double mk;
             private String nm;
             private int pn;
             private boolean preShow;
@@ -121,7 +121,7 @@ public class WaitMovieBean {
             private int recentShowDate;
             private int recentShowNum;
             private String rt;
-            private int sc;
+            private double sc;
             private String scm;
             private String showInfo;
             private int showNum;
@@ -232,11 +232,11 @@ public class WaitMovieBean {
                 this.localPubSt = localPubSt;
             }
 
-            public int getMk() {
+            public double getMk() {
                 return mk;
             }
 
-            public void setMk(int mk) {
+            public void setMk(double mk) {
                 this.mk = mk;
             }
 
@@ -304,11 +304,11 @@ public class WaitMovieBean {
                 this.rt = rt;
             }
 
-            public int getSc() {
+            public double getSc() {
                 return sc;
             }
 
-            public void setSc(int sc) {
+            public void setSc(double sc) {
                 this.sc = sc;
             }
